ENH: Subject hierarchy enhancements related to segmentations
Added concept of virtual subject hierarchy branches, which is a branch for which its parent is solely reaponsible, i.e. the child nodes do not correspond to actual MRML nodes, but to items contained in the parent node to enable per-item handling from subject hierarchy (e.g. segments in segmentations and fiducials in markups); Split volumes subject hierarchy plugin to scalar volumes and labelmaps plugin